LP64: fix byte->word sign-extension into 64-bit fields (UBSan audit)
Stood up a UBSan emu-g/emu build and swept the tests/lp64 suite, the
ref/limbobyexample corpus, and the inferno-lab corpus (incl. GUI apps under
Xvfb).  The one real LP64 bug class it surfaced: a uchar shifted << 24 promotes
to int, and for a high byte >= 0x80 (DMDIR 0x80, alpha 0xFF, ...) the result is
a negative int that sign-extends to 0xFFFFFFFF... when widened into a 64-bit
ulong/vlong.  Harmless on 32-bit (ulong was 4 bytes); wrong on LP64.

Fixed:
- include/styx.h, include/fcall.h: GBIT32/GBIT64 (the core 9P field-unpack
  macros) now assemble each byte as u32int, so mode/length/qid/time unpacks no
  longer sign-extend into bits 32..63; GBIT64 also zero-extends its low word.
- emu/port/dev.c (devdir), emu/port/devfs-posix.c, lib9/dirstat-{Nt,posix}.c:
  Dir.mode = (qid.type<<24)|... cast qid.type to ulong.
- libinterp/load.c: disw() assembles in u32int; the DEFL big-constant path and
  operand() (bytecode operand decoder) shift in u32int -- behavior-identical,
  removes the undefined-behavior left-shift-of-negative.

Verified: full suite 166/166, 9P round-trips clean under UBSan, the desktop and
the Mandelbrot (heavy per-pixel) render correctly.  Remaining UBSan findings
(pixel-assembly shifts, crypto/bignum byte-assembly, string hash, memmove(x,nil,0))
are benign 2's-complement UB with verified-correct results and were left to avoid
churn in hot paths.  ASan is incompatible with emu's pool allocator; UBSan +
Valgrind are the usable tools (see AGENTS_DEBUGGING.md "Sanitizer builds").

+++ emu/port/dev.c
@@ -36,7 +36,7 @@ devdir(Chan *c, Qid qid, char *n, long length, char *user, long perm, Dir *db)
 	db->qid = qid;
 	db->type = devtab[c->type]->dc;
 	db->dev = c->dev;
-	db->mode = perm | (qid.type << 24);
+	db->mode = perm | ((ulong)qid.type << 24);	/* LP64: avoid 0x80<<24 sign-extending into bits 32..63 */
 	db->atime = time(0);
 	db->mtime = kerndate;
 	db->length = length;
